misat completely functional 1.0 ready for exam
all info is already given, what do you want to read? Oh! You like the hobbit, noted:

In a hole in the ground there lived a hobbit. Not a nasty, dirty, wet hole, filled with the
ends of worms and an oozy smell, nor yet a dry, bare, sandy hole with nothing in it to sit
down on or to eat: it was a hobbit-hole, and that means comfort.